fix: import checkpoint instrumentation to avoid 64KB command limit
